Moved method getMedicialSpecializiation from class Commands to class MedicalSpecializiation.
